+++ Backend/src/main/java/com/virat/employeemanagementsystem/service/impl/EmployeeServiceImpl.java
@@ -1,24 +1,30 @@
 package com.virat.employeemanagementsystem.service.impl;
 
+import com.virat.employeemanagementsystem.common.response.PageResponse;
+import com.virat.employeemanagementsystem.dto.request.EmployeeFilterRequestDTO;
 import com.virat.employeemanagementsystem.dto.request.EmployeeRequestDTO;
 import com.virat.employeemanagementsystem.dto.response.EmployeeResponseDTO;
+import com.virat.employeemanagementsystem.entity.Department;
 import com.virat.employeemanagementsystem.entity.Employee;
 import com.virat.employeemanagementsystem.entity.Role;
+import com.virat.employeemanagementsystem.exception.DepartmentNotFoundException;
+import com.virat.employeemanagementsystem.exception.EmployeeDeletionException;
+import com.virat.employeemanagementsystem.exception.EmployeeNotFoundException;
 import com.virat.employeemanagementsystem.exception.InactiveRoleException;
 import com.virat.employeemanagementsystem.exception.RoleNotFoundException;
 import com.virat.employeemanagementsystem.mapper.EmployeeMapper;
 import com.virat.employeemanagementsystem.repository.DepartmentRepository;
 import com.virat.employeemanagementsystem.repository.EmployeeRepository;
 import com.virat.employeemanagementsystem.repository.RoleRepository;
+import com.virat.employeemanagementsystem.repository.UserRepository;
 import com.virat.employeemanagementsystem.service.EmployeeService;
-import org.springframework.transaction.annotation.Transactional;
+import com.virat.employeemanagementsystem.specification.EmployeeSpecification;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
-import com.virat.employeemanagementsystem.entity.Department;
-import com.virat.employeemanagementsystem.exception.DepartmentNotFoundException;
-import com.virat.employeemanagementsystem.exception.EmployeeNotFoundException;
-
-import java.util.List;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
@@ -32,19 +38,27 @@ public class EmployeeServiceImpl implements EmployeeService {
 
     private final RoleRepository roleRepository;
 
+    private final UserRepository userRepository;
+
     @Transactional
     @Override
     public EmployeeResponseDTO saveEmployee(EmployeeRequestDTO requestDTO) {
+
         Department department = findDepartmentById(
                 requestDTO.getDepartmentId()
         );
+
         Role role = findActiveRoleById(
                 requestDTO.getRoleId()
         );
+
         Employee employee = employeeMapper.toEntity(requestDTO);
+
         employee.setDepartment(department);
         employee.setRole(role);
+
         Employee savedEmployee = employeeRepository.save(employee);
+
         return employeeMapper.toResponseDTO(savedEmployee);
     }
 
@@ -57,11 +71,41 @@ public class EmployeeServiceImpl implements EmployeeService {
     }
 
     @Override
-    public List<EmployeeResponseDTO> getAllEmployees() {
+    public PageResponse<EmployeeResponseDTO> getAllEmployees(
+            Pageable pageable,
+            EmployeeFilterRequestDTO filterRequestDTO) {
 
-        List<Employee> employees = employeeRepository.findAll();
+        Specification<Employee> specification =
+                EmployeeSpecification.hasDepartment(filterRequestDTO.getDepartmentId());
 
-        return employeeMapper.toResponseDTOList(employees);
+        specification = specification.and(
+                EmployeeSpecification.hasRole(filterRequestDTO.getRoleId())
+        );
+
+        specification = specification
+                .and(EmployeeSpecification.hasDepartment(
+                        filterRequestDTO.getDepartmentId()
+                ))
+                .and(EmployeeSpecification.hasRole(
+                        filterRequestDTO.getRoleId()
+                ));
+
+        Page<Employee> employeePage =
+                employeeRepository.findAll(specification, pageable);
+
+        return PageResponse.<EmployeeResponseDTO>builder()
+                .content(
+                        employeeMapper.toResponseDTOList(
+                                employeePage.getContent()
+                        )
+                )
+                .page(employeePage.getNumber())
+                .size(employeePage.getSize())
+                .totalElements(employeePage.getTotalElements())
+                .totalPages(employeePage.getTotalPages())
+                .first(employeePage.isFirst())
+                .last(employeePage.isLast())
+                .build();
     }
 
     @Transactional
@@ -75,15 +119,17 @@ public class EmployeeServiceImpl implements EmployeeService {
         Department department = findDepartmentById(
                 requestDTO.getDepartmentId()
         );
+
         Role role = findActiveRoleById(
                 requestDTO.getRoleId()
         );
+
         employeeMapper.updateEntity(
                 requestDTO,
                 employee
         );
-        employee.setDepartment(department);
 
+        employee.setDepartment(department);
         employee.setRole(role);
 
         Employee updatedEmployee = employeeRepository.save(employee);
@@ -97,24 +143,40 @@ public class EmployeeServiceImpl implements EmployeeService {
 
         Employee employee = findEmployeeById(id);
 
+        if (userRepository.existsByEmployeeId(id)) {
+            throw new EmployeeDeletionException(
+                    "Cannot delete employee because a user account is associated with it."
+            );
+        }
+
         employeeRepository.delete(employee);
     }
 
     private Employee findEmployeeById(Long id) {
 
         return employeeRepository.findById(id)
-                .orElseThrow(() -> new EmployeeNotFoundException("Employee not found with id:"+id));
+                .orElseThrow(() ->
+                        new EmployeeNotFoundException(
+                                "Employee not found with id: " + id
+                        ));
     }
 
     private Department findDepartmentById(Long id) {
 
         return departmentRepository.findById(id)
-                .orElseThrow(() -> new DepartmentNotFoundException("Department not found with id:"+id));
+                .orElseThrow(() ->
+                        new DepartmentNotFoundException(
+                                "Department not found with id: " + id
+                        ));
     }
 
     private Role findActiveRoleById(Long id) {
+
         Role role = roleRepository.findById(id)
-                .orElseThrow(() -> new RoleNotFoundException("Role not found with id: "+id));
+                .orElseThrow(() ->
+                        new RoleNotFoundException(
+                                "Role not found with id: " + id
+                        ));
 
         if (!role.isActive()) {
             throw new InactiveRoleException(role.getId());
@@ -122,5 +184,4 @@ public class EmployeeServiceImpl implements EmployeeService {
 
         return role;
     }
-
 }
